requirements.txt
streamlit
numpy
scipy
scikit-image
imageio
Pillow